+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Fixed the `:mapping_immutable_input` shared examples so that it can deal with singleton inputs (nepalez)
 * Fixed typo in `WrongRule` exception message (nepalez)
 
+### Internal
+
+* Added the `Functions#identity` pure function (nepalez)
+
 [Compare v0.0.1...HEAD](https://github.com/nepalez/abstract_mapper/compare/v0.0.1...HEAD)
 
 ## v0.0.1 2015-07-04